+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/API/Types.hs
@@ -69,7 +69,7 @@ import Cardano.Ledger.PoolParams as X (
 import Cardano.Ledger.Shelley.BlockChain as X (bbHash)
 import Cardano.Ledger.Shelley.Genesis as X
 import Cardano.Ledger.Shelley.LedgerState as X (
-  AccountState (..),
+  ChainAccountState (..),
   DState (..),
   EpochState (..),
   EraCertState (..),