+++ libs/cardano-ledger-conformance/test/Test/Cardano/Ledger/Conformance/Imp/Ratify.hs
@@ -26,7 +26,7 @@ import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.Shelley.LedgerState (
   asTreasuryL,
   epochStateGovStateL,
-  esAccountStateL,
+  esChainAccountStateL,
   nesELL,
   nesEsL,
  )
@@ -61,7 +61,7 @@ spec = withImpInit @(LedgerSpec ConwayEra) $ describe "RATIFY" $ modifyImpInitPr
     govSt <- getsNES $ nesEsL . epochStateGovStateL
     let ratSt = getRatifyState govSt
     noConfidenceGAS <- getGovActionState noConfidence
-    treasury <- getsNES $ nesEsL . esAccountStateL . asTreasuryL
+    treasury <- getsNES $ nesEsL . esChainAccountStateL . asTreasuryL
     let
       execCtx =
         ConwayRatifyExecContext
@@ -128,7 +128,7 @@ spec = withImpInit @(LedgerSpec ConwayEra) $ describe "RATIFY" $ modifyImpInitPr
     constitutionGAS <- getGovActionState constitutionId
 
     logString "Testing conformance"
-    treasury <- getsNES $ nesEsL . esAccountStateL . asTreasuryL
+    treasury <- getsNES $ nesEsL . esChainAccountStateL . asTreasuryL
     let execCtx = ConwayRatifyExecContext treasury [constitutionGAS]
     ratEnv <- getRatifyEnv
     govSt <- getsNES $ nesEsL . epochStateGovStateL