+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.18.0.0
 
+* Rename `AccountState` to `ChainAccountState`. Rename accessor fields `asTreasury` and `asReserves` to `casTreasury` and `casReserves` respectively.
 * Added `sizeTxForFeeCalculation` to `EraTx` with a default implementation
 * Added `consumed` to `EraUTxO`
 * Removed `upgradeCertState`